revert loss

+++ monai/losses/dice.py
@@ -377,7 +377,7 @@ class GeneralizedDiceLoss(_Loss):
 
         denominator = ground_o + pred_o
 
-        w = self.w_func(ground_o.int())
+        w = self.w_func(ground_o.float())
         infs = torch.isinf(w)
         if self.batch:
             w[infs] = 0.0
@@ -623,11 +623,11 @@ class GeneralizedWassersteinDiceLoss(_Loss):
         Args:
             flat_target: the target tensor.
         """
-        alpha: torch.Tensor = torch.ones((flat_target.size(0), self.num_classes)).int().to(flat_target.device)
+        alpha: torch.Tensor = torch.ones((flat_target.size(0), self.num_classes)).float().to(flat_target.device)
         if self.alpha_mode == "GDL":  # GDL style
             # Define alpha like in the generalized dice loss
             # i.e. the inverse of the volume of each class.
-            one_hot_f = F.one_hot(flat_target, num_classes=self.num_classes).permute(0, 2, 1).int()
+            one_hot_f = F.one_hot(flat_target, num_classes=self.num_classes).permute(0, 2, 1).float()
             volumes = torch.sum(one_hot_f, dim=2)
             alpha = 1.0 / (volumes + 1.0)
         else:  # default, i.e. like in the original paper
